JavaScript Improvement
I had to update the JavaScript code for the positionHome() function for each site since I had improved on the home page originally. In the case of the About Us page, the JavaScript logic is getting super crowded. Because of this, a permanent solution will need to be done. It will most likely involve having images with built-in captions in the images and all the objects being centered in a div tag.  That's probably what is going to happen.