About Us Issue Fixed
I rewrote the entire HTML for the photo table. I replaced the table with a div tag that centers all of the images on the site. I still had to use a bit of JavaScript to manipulate some of the HTML during resizing, but the amount of code was less than when I used to have a table. Also, I used CSS to write a few media queries that could help out with resizing the pictures.