Removing Unnecessary Code and Modifying Store Page
In this case, the change was that I had to modify parts of the Store page by adding a media query and updating the function that positions the Home button. However, the code that I wanted to use was in contact-us, and it had code that was commented out. Because of this, I removed the comments, copied the code, and pasted it on the other site. I think we are good to go!