Updating Developer Info and Removing REMs.
I had to update the way the developer contact info is displayed due to the fact that it did not look good when it popped up on mobile devices. Because of this, I had to improve that capability. Also, I removed several commented-out pieces of code that I didn't need anymore. It would take too much space. There were other pieces of code that I removed too, due to the fact that it wasn't needed anymore.