More Updating to Developer Information and Removing Operating System Check Calls
I had to update the way the developer information was shown on the website. I also had to remove all of the calls that I used for checking the operating system name of a particular device. I'll only keep it for checking the kind of operating system if there has been a bug that has occurred on one operating system, and I have to fix it.