New Images, New Code!
I had to write some code that could show each of these cute rabbits. This was just a start, however. I may have to deal with CSS and image sizes. Most likely, media queries can help with this, due to the fact that all of the images that were taken with the camera are all the same size. I can just set them at the same rate, which can be very helpful.