Rabbit Image Positioning Improved
On the Meet Our Bunnies page, I had to improve the positions of the bunnies. Of course, there is a lot more to do in this case, but this is a good move in the right direction! I had to use some JavaScript to help with repositioning for the mobile devices (we know how many people use those!).