Bug Fix
There was a bug in the way that clicking on the rabbits worked. Because of this, I found a fix. Turns out that, by updating the HTML for the rabbit pictures, I was making the pictures not clickable. By turning it off, and putting the figcaptions in the right spot, it was all fixed. :smiley: