Adding Charts and Citing Sources
I added a chart to one of the sections where we are introducing a rabbit. Because of this, I wrote some extra code to help with citing sources. Most of the images that we use are mostly produced by the photographer/video editor (except for the Facebook and YouTube logos).