More Rabbits, More Information, New Code
I added two more rabbits to the Meet Our Bunnies page. I got the chance to write new code that allows there to be links between different rabbits. This may need to be improved on as more rabbits are added in, but it is a start. I had to remove the ability to scroll automatically back to where you are at, but overall, the results look great! It was nice to add a bit more JavaScript code into the website. That was a bit of a treat.