More Updates
I have made a few changes to this particular page's JavaScript code. First of all, I found a way to automatically scroll the user to the place where they last were looking at rabbits before they clicked on an external link. I also started to see if implementing an array would be useful or not. Originally, the plain array itself was not working, so I have begun to experiment with using a stack to help out with this. This is still a work in progress, but hey, things are looking up!