New Updates
After a bit of investigation, I realized that I was pushing the rabbit I was currently on twice by calling them in two separate functions. Because of this, I had to remove that. I also was able to use the stack data structure as a replacement for the previousRabbit variable.

I later on replaced the stack with an array implementation to save lines of code. Everything seems to be working smoothly. I may remove the console.log lines later on.