REM removal
I removed most of the REMs from the site. The next part of the project is to write some media queries that can be used as a replacement for the JavaScript code that runs on the positionHome. That way, the only task that the JavaScript code would be working on is positioning the rabbit images.