Update index.html
I removed the extra new line. I experimented with media queries to see if they would work in theory, but it took away the flow of fitting in the screen. So, I'm sticking with the JavaScript implementation of the solution. Of course, I don't use the JavaScript implementation for my other websites. I just use the CSS media query method. However, I think it is worth keeping the JavaScript implementation just to keep the flow. Also, it shows one of many ways to accomplish the same thing: making everything fit on every screen.