Fixing Popup Issues
I had to fix a couple of issues involving the popup that would show anytime someone clicked on a rabbit. I wanted to make sure that the information fit on the screen relatively well, so I added a couple of media queries to the CSS to ensure the popup fit on the screen properly. I also had to remove a media query that was contradicting another, thus solving another issue relating to the UI.