New Rabbits
It has been a while since I have added more rabbits to the site. Turns out, some of the images that I used are not the right ones. Because of this, I had to remove a couple of the images. However, I was able to add two new ones, so that is good.

I think the next thing that I will be doing to this site is starting to work on putting rabbits that we are selling. I have to work on making this functional.

I also have to start working on the YouTube channel for this organization. I think the first thing that we need to do is just create videos of rabbits that we are selling and showing them off. After that, we can do more complex things, such as doing rabbit biology facts and things like that.

Alright, enough of my rambling. This has been fun. I do miss writing all of the code for the site, though. Nowadays, the code changes are minor. In some cases, I do major changes to the code, such as if there is something that I have to change in the JavaScript, but otherwise it is just editing HTML. That's fine, though. It is still fun work to do.