Removed Developer Contact Form
I commented out the developer contact form in order to look a bit more competent in my work as a web developer. This was something that a designer friend of mine explained to me. Pretty soon, I'll be building a website for him as well. He'll be doing the design portion, however.

Hmm. I wonder if I could use Bootstrap that time? So far, I have had to write my own code in order to make the website responsive. However, turns out that using Bootstrap may be better.